+++ game.js
@@ -78,6 +78,7 @@ var Game = (function() {
         self.researchUI.update(delta);
         self.ascensionUI.update(delta);
         self.structureUI.update(delta);
+        self.worldCycleUI.update(delta);
 
         self.updateAutoSave(delta);
 
@@ -162,6 +163,7 @@ var Game = (function() {
         this.research.save(data);
         this.ascension.save(data);
         this.structures.save(data);
+        this.worldCycle.save(data);
         this.interstellar.save(data);
         this.stargaze.save(data);
         this.updates.save(data);
@@ -192,6 +194,7 @@ var Game = (function() {
             this.research.load(data);
             this.ascension.load(data);
             this.structures.load(data);
+            this.worldCycle.load(data);
             this.stargaze.load(data);
             this.tech.load(data);
             this.interstellar.load(data); 
@@ -281,6 +284,7 @@ var Game = (function() {
         self.research.initialise();
         self.ascension.initialise();
         self.structures.initialise();
+        self.worldCycle.initialise();
 
         // Now load
         self.load();
@@ -296,6 +300,7 @@ var Game = (function() {
         self.researchUI.initialise();
         self.ascensionUI.initialise();
         self.structureUI.initialise();
+        self.worldCycleUI.initialise();
 
         for(var i = 0; i < self.uiComponents.length; i++) {
             self.uiComponents[i].initialise();